Export comment functions from typemap.Registry
Closes #1
