+++ typemap/typemap.go
@@ -15,7 +15,6 @@ package typemap
 
 import (
 	"github.com/golang/protobuf/protoc-gen-go/descriptor"
-	"github.com/pkg/errors"
 )
 
 // Registry is an index for finding the definitions of messages, given a message
@@ -53,33 +52,19 @@ func New(files []*descriptor.FileDescriptorProto) *Registry {
 	return r
 }
 
+// Deprecated: use typemap.FileComments instead.
 func (r *Registry) FileComments(file *descriptor.FileDescriptorProto) (DefinitionComments, error) {
-	return commentsAtPath([]int32{packagePath}, file), nil
+	return FileComments(file)
 }
 
+// Deprecated: use typemap.ServiceComments instead.
 func (r *Registry) ServiceComments(file *descriptor.FileDescriptorProto, svc *descriptor.ServiceDescriptorProto) (DefinitionComments, error) {
-	for i, s := range file.Service {
-		if s == svc {
-			path := []int32{servicePath, int32(i)}
-			return commentsAtPath(path, file), nil
-		}
-	}
-	return DefinitionComments{}, errors.Errorf("service not found in file")
+	return ServiceComments(file, svc)
 }
 
+// Deprecated: use typemap.MethodComments instead.
 func (r *Registry) MethodComments(file *descriptor.FileDescriptorProto, svc *descriptor.ServiceDescriptorProto, method *descriptor.MethodDescriptorProto) (DefinitionComments, error) {
-	for i, s := range file.Service {
-		if s == svc {
-			path := []int32{servicePath, int32(i)}
-			for j, m := range s.Method {
-				if m == method {
-					path = append(path, serviceMethodPath, int32(j))
-					return commentsAtPath(path, file), nil
-				}
-			}
-		}
-	}
-	return DefinitionComments{}, errors.Errorf("service not found in file")
+	return MethodComments(file, svc, method)
 }
 
 func (r *Registry) MethodInputDefinition(method *descriptor.MethodDescriptorProto) *MessageDefinition {
@@ -204,90 +189,6 @@ func messageDefsForFile(f *descriptor.FileDescriptorProto, filesByName map[strin
 	}
 
 	return byProtoName
-}
-
-// DefinitionComments contains the comments surrounding a definition in a
-// protobuf file.
-//
-// These follow the rules described by protobuf:
-//
-// A series of line comments appearing on consecutive lines, with no other
-// tokens appearing on those lines, will be treated as a single comment.
-//
-// leading_detached_comments will keep paragraphs of comments that appear
-// before (but not connected to) the current element. Each paragraph,
-// separated by empty lines, will be one comment element in the repeated
-// field.
-//
-// Only the comment content is provided; comment markers (e.g. //) are
-// stripped out.  For block comments, leading whitespace and an asterisk
-// will be stripped from the beginning of each line other than the first.
-// Newlines are included in the output.
-//
-// Examples:
-//
-//   optional int32 foo = 1;  // Comment attached to foo.
-//   // Comment attached to bar.
-//   optional int32 bar = 2;
-//
-//   optional string baz = 3;
-//   // Comment attached to baz.
-//   // Another line attached to baz.
-//
-//   // Comment attached to qux.
-//   //
-//   // Another line attached to qux.
-//   optional double qux = 4;
-//
-//   // Detached comment for corge. This is not leading or trailing comments
-//   // to qux or corge because there are blank lines separating it from
-//   // both.
-//
-//   // Detached comment for corge paragraph 2.
-//
-//   optional string corge = 5;
-//   /* Block comment attached
-//    * to corge.  Leading asterisks
-//    * will be removed. */
-//   /* Block comment attached to
-//    * grault. */
-//   optional int32 grault = 6;
-//
-//   // ignored detached comments.
-type DefinitionComments struct {
-	Leading         string
-	Trailing        string
-	LeadingDetached []string
-}
-
-func commentsAtPath(path []int32, sourceFile *descriptor.FileDescriptorProto) DefinitionComments {
-	if sourceFile.SourceCodeInfo == nil {
-		// The compiler didn't provide us with comments.
-		return DefinitionComments{}
-	}
-
-	for _, loc := range sourceFile.SourceCodeInfo.Location {
-		if pathEqual(path, loc.Path) {
-			return DefinitionComments{
-				Leading:         loc.GetLeadingComments(),
-				LeadingDetached: loc.GetLeadingDetachedComments(),
-				Trailing:        loc.GetTrailingComments(),
-			}
-		}
-	}
-	return DefinitionComments{}
-}
-
-func pathEqual(path1, path2 []int32) bool {
-	if len(path1) != len(path2) {
-		return false
-	}
-	for i, v := range path1 {
-		if path2[i] != v {
-			return false
-		}
-	}
-	return true
 }
 
 const (
